+++ library-collection/src/androidTest/java/universum/studios/android/preference/ArrayPreferenceTest.java
@@ -34,7 +34,7 @@ import universum.studios.android.test.PreferencesTest;
 import static junit.framework.Assert.assertEquals;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.core.IsNull.nullValue;
+import static org.hamcrest.Matchers.nullValue;
 
 /**
  * @author Martin Albedinsky