+++ library-collection/src/main/java/universum/studios/android/preference/ListPreference.java
@@ -24,17 +24,20 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.annotation.StringRes;
 
-import java.util.Arrays;
 import java.util.List;
 
 /**
+ * <b>This class has been deprecated and will be removed in the next release.</b>
+ * <p>
  * A {@link SharedPreference} implementation that may be used to persist a {@link List} of values
  * via {@link SharedPreferences}.
  *
  * @param <T> Type of items within a list of which values should be persisted.
  * @author Martin Albedinsky
  * @see ArrayPreference
+ * @deprecated Use {@link CollectionPreference} instead.
  */
+@Deprecated
 public final class ListPreference<T> extends SharedPreference<List<T>> {
 
 	/*
@@ -42,9 +45,8 @@ public final class ListPreference<T> extends SharedPreference<List<T>> {
 	 */
 
 	/**
-	 * Class of components stored within a list managed by this preference.
 	 */
-	private final Class<T> mComponentType;
+	private final CollectionPreference<T> mDelegate;
 
 	/*
 	 * Constructors ================================================================================
@@ -59,7 +61,7 @@ public final class ListPreference<T> extends SharedPreference<List<T>> {
 	 */
 	public ListPreference(@NonNull final String key, @NonNull final Class<T> componentType, @Nullable final List<T> defValue) {
 		super(key, defValue);
-		this.mComponentType = componentType;
+		this.mDelegate = new CollectionPreference<>(key, componentType, defValue);
 	}
 
 	/**
@@ -75,7 +77,7 @@ public final class ListPreference<T> extends SharedPreference<List<T>> {
 	@Deprecated
 	public ListPreference(@StringRes int keyResId, @NonNull Class<T> componentType, @Nullable List<T> defValue) {
 		super(keyResId, defValue);
-		this.mComponentType = componentType;
+		this.mDelegate = new CollectionPreference<>(keyResId, componentType, defValue);
 	}
 
 	/*
@@ -88,7 +90,7 @@ public final class ListPreference<T> extends SharedPreference<List<T>> {
 	@Override
 	@CheckResult
 	protected boolean onPutIntoPreferences(@NonNull final SharedPreferences preferences) {
-		return putIntoPreferences(preferences, mKey, mValue, mComponentType);
+		return mDelegate.onPutIntoPreferences(preferences);
 	}
 
 	/**
@@ -104,22 +106,7 @@ public final class ListPreference<T> extends SharedPreference<List<T>> {
 	@CheckResult
 	@SuppressWarnings("unchecked")
 	public static <T> boolean putIntoPreferences(@NonNull final SharedPreferences preferences, @NonNull final String key, @Nullable final List<T> list, @NonNull final Class<T> componentType) {
-		final SharedPreferences.Editor editor = preferences.edit();
-		if (list == null) {
-			editor.putString(key, null);
-		} else {
-			final T[] array = (T[]) ArrayPreference.createArrayInSize(componentType, list.size());
-			if (array == null) {
-				final String componentName = componentType.getSimpleName();
-				throw new IllegalArgumentException(
-						"Failed to put list of(" + componentName + ") into shared preferences. " +
-								"Only lists of primitive types or theirs boxed representations including String are supported."
-				);
-			}
-			list.toArray(array);
-			return ArrayPreference.putIntoPreferences(preferences, key, array);
-		}
-		return editor.commit();
+		return CollectionPreference.putIntoPreferences(preferences, key, list, componentType);
 	}
 
 	/**
@@ -151,28 +138,6 @@ public final class ListPreference<T> extends SharedPreference<List<T>> {
 	 */
 	@SuppressWarnings("unchecked")
 	public static <T> List<T> getFromPreferences(@NonNull final SharedPreferences preferences, @NonNull final String key, @Nullable final List<T> defValue) {
-		final String value = preferences.getString(key, null);
-		if (value != null) {
-			try {
-				final T[] array = ArrayPreference.getFromPreferences(preferences, key, null);
-				return Arrays.asList(array);
-			} catch (ClassCastException e) {
-				final String arrayValue = ArrayPreference.extractArrayValueFromPreferenceValue(value);
-				throw new ClassCastException(
-						"Cannot obtain a list for the key(" + key + ") from shared preferences. " +
-								"Value(" + arrayValue + ") is not a list!"
-				);
-			} catch (IllegalArgumentException e) {
-				throw new IllegalArgumentException(
-						"Failed to obtain list for the key(" + key + ") from shared preferences. " +
-								"Only lists of primitive types or theirs boxed representations including String are supported."
-				);
-			} catch (IllegalStateException e) {
-				throw new IllegalStateException(
-						"Trying to obtain a list for the key(" + key + ") from shared preferences not saved by the Preferences library."
-				);
-			}
-		}
-		return defValue;
+		return (List<T>) CollectionPreference.getFromPreferences(preferences, key, defValue);
 	}
 }